requirements.txt
pyyaml
orjson
json5
loguru
aiogram>=3.22.0
aiohttp>=3.9.4
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
